+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Title from "./Title";
+import Title from "./components/Title";
 
 const App = () => <Title text="hello typescript" />;
